# Conflicts:
#	VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.AutomationService.Contracts/ServiceContracts.autogen.cs
#	VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.AutomationService.Contracts/Services.autogen.cs
#	VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.AutomationService/Controllers/IdentityController.cs

+++ VerticalWarehouses/Common/Ferretto.VW.CommonUtils/Messages/Data/DrawerOperationMessageData.cs
@@ -1,14 +1,15 @@
 ﻿using Ferretto.VW.CommonUtils.Messages.Enumerations;
 using Ferretto.VW.CommonUtils.Messages.Interfaces;
-// ReSharper disable ArrangeThisQualifier
 
+// ReSharper disable ArrangeThisQualifier
 namespace Ferretto.VW.CommonUtils.Messages.Data
 {
     public class DrawerOperationMessageData : IDrawerOperationMessageData
     {
         #region Constructors
 
-        public DrawerOperationMessageData(DrawerOperation operation,
+        public DrawerOperationMessageData(
+            DrawerOperation operation,
             MessageVerbosity verbosity = MessageVerbosity.Debug)
         {
             this.Operation = operation;
@@ -21,9 +22,9 @@ namespace Ferretto.VW.CommonUtils.Messages.Data
 
         public LoadingUnitDestination Destination { get; set; }
 
-        public decimal DestinationHorizontalPosition { get; set; }
+        public double DestinationHorizontalPosition { get; set; }
 
-        public decimal DestinationVerticalPosition { get; set; }
+        public double DestinationVerticalPosition { get; set; }
 
         public bool IsDestinationPositive { get; set; }
 
@@ -33,9 +34,9 @@ namespace Ferretto.VW.CommonUtils.Messages.Data
 
         public LoadingUnitDestination Source { get; set; }
 
-        public decimal SourceHorizontalPosition { get; set; }
+        public double SourceHorizontalPosition { get; set; }
 
-        public decimal SourceVerticalPosition { get; set; }
+        public double SourceVerticalPosition { get; set; }
 
         public MessageVerbosity Verbosity { get; }
