+++ VerticalWarehouses/Common/Ferretto.VW.CommonUtils/Messages/Data/WeightAcquisitionCommandMessageData.cs
@@ -10,13 +10,13 @@ namespace Ferretto.VW.CommonUtils.Messages.Data
 
         public CommandAction CommandAction { get; set; }
 
-        public decimal Displacement { get; set; }
+        public double Displacement { get; set; }
 
-        public decimal InitialPosition { get; set; }
+        public double InitialPosition { get; set; }
 
         public TimeSpan InPlaceSamplingDuration { get; set; }
 
-        public decimal NetWeight { get; set; }
+        public double NetWeight { get; set; }
 
         public MessageVerbosity Verbosity { get; set; }
